[refactor] 이미지 추가 CollectionView 관련 클래스 분리 (#9)
