+++ SpaceCapsule/SpaceCapsule/Extensions/UIImage+.swift
@@ -24,4 +24,5 @@ extension UIImage {
     static let chevronRight = UIImage(systemName: "chevron.right")
     static let close = UIImage(systemName: "xmark")
     static let sort = UIImage(systemName: "arrow.up.arrow.down")
+    static let lock = UIImage(systemName: "lock.fill")
 }
