+++ SpaceCapsule/SpaceCapsule/Scene/TabBar/CapsuleAdd/CapsuleCreate/CapsuleCreateView.swift
@@ -8,82 +8,6 @@
 import SnapKit
 import UIKit
 
-final class CapsuleCreateCollectionViewCell: UICollectionViewCell {
-    let imageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.contentMode = .scaleToFill
-
-        return imageView
-    }()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        addSubViews()
-        makeConstraints()
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func configure(item: String) {
-        imageView.image = UIImage(named: item)
-    }
-
-    private func addSubViews() {
-        addSubview(imageView)
-    }
-
-    private func makeConstraints() {
-        imageView.snp.makeConstraints {
-            $0.top.equalToSuperview()
-            $0.bottom.equalToSuperview()
-            $0.leading.equalToSuperview()
-            $0.trailing.equalToSuperview()
-        }
-    }
-}
-
-final class CapsuleCreateCollectionView: UICollectionView {
-    required init(frame: CGRect) {
-        super.init(frame: frame, collectionViewLayout: CapsuleCreateCollectionView.layout())
-
-        register(
-            CapsuleCreateCollectionViewCell.self,
-            forCellWithReuseIdentifier: CapsuleCreateCollectionViewCell.identifier
-        )
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    private static func layout() -> UICollectionViewCompositionalLayout {
-        let itemSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalHeight(1),
-            heightDimension: .fractionalHeight(1)
-        )
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-
-        let groupSize = NSCollectionLayoutSize(
-            widthDimension: .fractionalHeight(1),
-            heightDimension: .fractionalHeight(1)
-        )
-        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
-
-        let section = NSCollectionLayoutSection(group: group)
-        section.orthogonalScrollingBehavior = .groupPagingCentered
-        section.interGroupSpacing = 20
-
-        let layout = UICollectionViewCompositionalLayout(section: section)
-
-        return layout
-    }
-}
-
 final class CapsuleCreateView: UIView, BaseView {
     private let mainStackView = UIStackView()
     private let inputStackView: UIStackView = {
@@ -100,7 +24,7 @@ final class CapsuleCreateView: UIView, BaseView {
 
     let titleTextField = ThemeTextField()
 
-    lazy var imageCollectionView = CapsuleCreateCollectionView(frame: frame)
+    lazy var imageCollectionView = AddImageCollectionView(frame: frame)
     private let sampleImages: [UIImage?] = [.logoWithBG, .logoWithText, .logo]
 
     override init(frame: CGRect) {
