+++ SpaceCapsule/SpaceCapsule/Scene/TabBar/CapsuleAdd/CapsuleCreate/Components/AddImageCollectionView.swift
@@ -1,0 +1,63 @@
+//
+//  AddImageCollectionView.swift
+//  SpaceCapsule
+//
+//  Created by 장재훈 on 2022/11/19.
+//
+
+import UIKit
+
+final class AddImageCollectionView: UICollectionView {
+    required init(frame: CGRect) {
+        super.init(frame: frame, collectionViewLayout: AddImageCollectionView.layout())
+
+        register(
+            AddImageCollectionViewCell.self,
+            forCellWithReuseIdentifier: AddImageCollectionViewCell.identifier
+        )
+
+        register(
+            AddImageCollectionViewFooter.self,
+            forSupplementaryViewOfKind: UICollectionView.elementKindSectionFooter,
+            withReuseIdentifier: AddImageCollectionViewFooter.identifier
+        )
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    private static func layout() -> UICollectionViewCompositionalLayout {
+        let itemSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalHeight(1),
+            heightDimension: .fractionalHeight(1)
+        )
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+
+        let groupSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalHeight(1),
+            heightDimension: .fractionalHeight(1)
+        )
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+
+        let footerSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalHeight(1),
+            heightDimension: .fractionalHeight(1)
+        )
+        let footer = NSCollectionLayoutBoundarySupplementaryItem(
+            layoutSize: footerSize,
+            elementKind: UICollectionView.elementKindSectionFooter,
+            alignment: .bottom
+        )
+
+        let section = NSCollectionLayoutSection(group: group)
+        section.orthogonalScrollingBehavior = .groupPagingCentered
+        section.interGroupSpacing = 20
+        section.boundarySupplementaryItems = [footer]
+
+        let layout = UICollectionViewCompositionalLayout(section: section)
+
+        return layout
+    }
+}
